Issue #48 Attempt to move pause movement to update_state to avoid delays -e Original/build commit ID: 9ddf1d9d1338c80b01bd8fba2f156f5a81fe1ac3

+++ game_state.c
@@ -70,7 +70,7 @@ typedef struct {
     // Inventory items
     UINT8 inventory[7U];
     
-    UINT8 last_movement_time;
+    unsigned int last_movement_time;
 } game_state_t;
